requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-admin-rangefilter==0.13.0
django-tables2==2.7.0
djangorestframework==3.15.1
fpdf==1.7.2
idna==3.7
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
